dao,page,manage
-- modify dao
-- add milktea page, game page, manage tools page.

+++ Program/web/src/dev/boom/core/BoomSession.java
@@ -5,18 +5,18 @@ import java.io.Serializable;
 public class BoomSession implements Serializable{
 
 	private static final long serialVersionUID = 1L;
-	private int id;
+	private long id;
 
-	public BoomSession(int id) {
+	public BoomSession(long id) {
 		super();
 		this.id = id;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
